resolved conflicts

+++ indra/newview/llagentcamera.cpp
@@ -189,8 +189,10 @@ LLAgentCamera::LLAgentCamera() :
     mPanRightKey(0.f),
     mPanInKey(0.f),
     mPanOutKey(0.f),
+// <FS:Chanayane> Camera roll (from Alchemy)
     mRollLeftKey(0.f),
     mRollRightKey(0.f),
+// </FS:Chanayane>
     mPointAtObject(NULL)
 {
     mFollowCam.setMaxCameraDistantFromSubject( MAX_CAMERA_DISTANCE_FROM_AGENT );
@@ -201,7 +203,9 @@ LLAgentCamera::LLAgentCamera() :
 
     resetPanDiff();
     resetOrbitDiff();
+// <FS:Chanayane> Camera roll (from Alchemy)
 	resetCameraRoll();
+// </FS:Chanayane>
 }
 
 // Requires gSavedSettings to be initialized.
@@ -383,7 +387,9 @@ void LLAgentCamera::resetView(BOOL reset_camera, BOOL change_camera, BOOL moveme
         setFocusOnAvatar(TRUE, ANIMATE);
 
         mCameraFOVZoomFactor = 0.f;
+// <FS:Chanayane> Camera roll (from Alchemy)
 		resetCameraRoll();
+// </FS:Chanayane>
     }
     resetPanDiff();
     resetOrbitDiff();
@@ -940,18 +946,20 @@ void LLAgentCamera::cameraOrbitOver(const F32 angle)
     }
 }
 
+// <FS:Chanayane> Camera roll (from Alchemy)
 //-----------------------------------------------------------------------------
 // cameraRollOver()
 //-----------------------------------------------------------------------------
 void LLAgentCamera::cameraRollOver(const F32 angle)
 {
-    mRollAngle += fmod(angle, F_TWO_PI);
+    mRollAngle += fmodf(angle, F_TWO_PI);
 }
 
 void LLAgentCamera::resetCameraRoll()
 {
     mRollAngle = 0.f;
 }
+// </FS:Chanayane>
 void LLAgentCamera::resetCameraOrbit()
 {
     LLVector3 camera_offset_unit(mCameraFocusOffsetTarget);
@@ -965,7 +973,9 @@ void LLAgentCamera::resetCameraOrbit()
 
     cameraZoomIn(1.f);
     resetOrbitDiff();
+// <FS:Chanayane> Camera roll (from Alchemy)
 	resetCameraRoll();
+// </FS:Chanayane>
 }
 
 void LLAgentCamera::resetOrbitDiff()
@@ -1372,7 +1382,9 @@ void LLAgentCamera::updateCamera()
     const F32 ORBIT_OVER_RATE = 90.f * DEG_TO_RAD;          // radians per second
     const F32 ORBIT_AROUND_RATE = 90.f * DEG_TO_RAD;        // radians per second
     const F32 PAN_RATE = 5.f;                               // meters per second
+// <FS:Chanayane> Camera roll (from Alchemy)
     const F32 ROLL_RATE = 45.f * DEG_TO_RAD;                // radians per second
+// </FS:Chanayane>
 
     if (gAgentCamera.getOrbitUpKey() || gAgentCamera.getOrbitDownKey())
     {
@@ -1414,11 +1426,14 @@ void LLAgentCamera::updateCamera()
         cameraPanUp(input_rate * PAN_RATE / gFPSClamped );
     }
 
+// <FS:Chanayane> Camera roll (from Alchemy)
     if (getRollLeftKey() || getRollRightKey())
     {
         F32 input_rate = getRollRightKey() - getRollLeftKey();
         cameraRollOver(input_rate * ROLL_RATE / gFPSClamped);
     }
+// </FS:Chanayane>
+
     // Clear camera keyboard keys.
     gAgentCamera.clearOrbitKeys();
     gAgentCamera.clearPanKeys();
@@ -1663,6 +1678,7 @@ void LLAgentCamera::updateCamera()
         torso_joint->setScale(torso_scale);
         chest_joint->setScale(chest_scale);
     }
+// <FS:Chanayane> Camera roll (from Alchemy)
     //     We have do this at the very end to make sure it takes all previous calculations into
     //     account and then applies our roll on top of it, besides it wouldn't even work otherwise.
     LLQuaternion rot_quat = LLViewerCamera::getInstance()->getQuaternion();
@@ -1675,6 +1691,7 @@ void LLAgentCamera::updateCamera()
     LLViewerCamera::getInstance()->mYAxis = LLVector3(mat.mMatrix[1]);
     LLViewerCamera::getInstance()->mZAxis = LLVector3(mat.mMatrix[2]);
 }
+// </FS:Chanayane>
 
 void LLAgentCamera::updateLastCamera()
 {
@@ -2813,7 +2830,9 @@ void LLAgentCamera::switchCameraPreset(ECameraPreset preset)
 
     resetPanDiff();
     resetOrbitDiff();
+// <FS:Chanayane> Camera roll (from Alchemy)
 	resetCameraRoll();
+// </FS:Chanayane>
 
     gSavedSettings.setU32("CameraPresetType", mCameraPreset);
 }
@@ -3318,8 +3337,10 @@ void LLAgentCamera::clearOrbitKeys()
     mOrbitDownKey       = 0.f;
     mOrbitInKey         = 0.f;
     mOrbitOutKey        = 0.f;
+// <FS:Chanayane> Camera roll (from Alchemy)
     mRollLeftKey        = 0.f;
     mRollRightKey       = 0.f;
+// </FS:Chanayane>
 }
 
 void LLAgentCamera::clearPanKeys()
@@ -3350,7 +3371,9 @@ void LLAgentCamera::storeCameraPosition()
     // flycam  mode and not repositioned after
     LLVector3d forward = LLVector3d(1.0, 0.0, 0.0) * LLViewerCamera::getInstance()->getQuaternion() + getCameraPositionGlobal();
     gSavedPerAccountSettings.setVector3d("FSStoredCameraFocus", forward);
+// <FS:Chanayane> Camera roll (from Alchemy)
 	gSavedPerAccountSettings.setF32("ALStoredCameraRoll", mRollAngle);
+// </FS:Chanayane>
 
     LLUUID stored_camera_focus_object_id = LLUUID::null;
     if (mFocusObject)
@@ -3364,7 +3387,9 @@ void LLAgentCamera::loadCameraPosition()
 {
     LLVector3d stored_camera_pos = gSavedPerAccountSettings.getVector3d("FSStoredCameraPos");
     LLVector3d stored_camera_focus = gSavedPerAccountSettings.getVector3d("FSStoredCameraFocus");
+// <FS:Chanayane> Camera roll (from Alchemy)
     F32 stored_camera_roll = gSavedPerAccountSettings.getF32("ALStoredCameraRoll");
+// </FS:Chanayane>
     LLUUID stored_camera_focus_object_id = LLUUID(gSavedPerAccountSettings.getString("FSStoredCameraFocusObjectId"));
 
     F32 renderFarClip = gSavedSettings.getF32("RenderFarClip");
@@ -3393,7 +3418,9 @@ void LLAgentCamera::loadCameraPosition()
 
     unlockView();
     setCameraPosAndFocusGlobal(stored_camera_pos, stored_camera_focus, stored_camera_focus_object_id);
+// <FS:Chanayane> Camera roll (from Alchemy)
 	mRollAngle = stored_camera_roll;
+// </FS:Chanayane>
 }
 // </FS:Ansariel> FIRE-7758: Save/load camera position feature
